chore: Cleanup TS types
