+++ src/react/entry-client.ts
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom'
 import { HelmetProvider } from 'react-helmet-async'
 import { withoutSuffix } from '../utils/route'
 import { createRouter } from './utils'
-export { ClientOnly } from './components.js'
+export { ClientOnly } from './components'
 import type { ClientHandler } from './types'
 
 export const viteSSR: ClientHandler = async function (
